Move --retrieve-cplex-mip-duals capability  from hawaii.switch_patch to core model.
